backend/requirements.txt
# FastAPI and Web Framework
fastapi[standard]>=0.115.7,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
pydantic-core>=2.27.0,<3.0.0
psycopg2

# Database
sqlalchemy>=2.0.23,<2.1.0
alembic>=1.13.1,<2.0.0
asyncpg>=0.29.0,<0.30.0
redis>=5.0.1,<6.0.0

# Authentication and Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.18,<1.0.0

# HTTP Client
httpx>=0.28.1,<1.0.0
aiohttp>=3.10.0,<4.0.0

# AI/ML
google-genai>=1.18.0,<2.0.0
scikit-learn>=1.3.2,<2.0.0
numpy>=1.24.4,<2.0.0
pandas>=2.1.4,<3.0.0
pillow>=10.1.0,<11.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
email-validator>=2.1.0,<3.0.0
psutil>=5.9.6,<6.0.0

# Testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-mock>=3.12.0,<4.0.0

# Development
black>=23.11.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<8.0.0
mypy>=1.7.1,<2.0.0

# Monitoring
prometheus-fastapi-instrumentator>=6.1.0,<7.0.0
prometheus-client>=0.19.0,<1.0.0
discord-webhook>=1.3.0,<2.0.0
